intermediate refactoring 3
